Break out table model
